backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.10
pydantic==2.9.2
pydantic-settings==2.6.0
python-multipart==0.0.12
pyyaml==6.0.3
websockets==13.1
bcrypt==5.0.0
python-jose[cryptography]==3.3.0
email-validator==2.2.0
